+++ src/components/nav/NavBarLogo.tsx
@@ -1,17 +1,14 @@
-import { Image } from '@chakra-ui/image'
 import { Box, BoxProps } from '@chakra-ui/layout'
 import { useBreakpoint } from '@chakra-ui/media-query'
-import { useColorMode } from '@chakra-ui/system'
 import { createUseStyles } from 'react-jss'
 import { Link } from 'react-router-dom'
 
-import LogoSmall from '../../assets/logo-brand.svg'
-import LogoDark from '../../assets/logo-dark.svg'
-import { LogoNameBrand } from '../../constants'
+import { GeyserLogoNameIcon } from '../icons/svg'
+import { GeyserLogoIcon } from '../icons/svg/GeyserLogoIcon'
 
 type Props = BoxProps & {
   className?: string
-  imageClassName?: string
+  color?: string
   full?: boolean
   small?: boolean
 }
@@ -26,37 +23,33 @@ const useStyles = createUseStyles({
 
 export const NavBarLogo = ({
   className,
-  imageClassName,
   full,
   small,
+  color,
   ...rest
 }: Props) => {
   const classes = useStyles()
-  const { colorMode } = useColorMode()
 
   const isLg = useBreakpoint('lg')
 
   const useFullOne = (isLg || full) && !small
 
-  const imageToUse =
-    colorMode === 'light'
-      ? useFullOne
-        ? LogoNameBrand
-        : LogoSmall
-      : useFullOne
-      ? LogoNameBrand
-      : LogoDark
-
   return (
     <Box {...rest} className={`${classes.container} ${className}`}>
       <Link to="/">
-        <Image
-          className={imageClassName}
-          height="32px"
-          src={imageToUse}
-          alt="geyser logo image"
-          objectFit="contain"
-        />
+        {useFullOne ? (
+          <GeyserLogoNameIcon
+            height="32px"
+            width="100%"
+            color={color || 'primary.400'}
+          />
+        ) : (
+          <GeyserLogoIcon
+            height="32px"
+            width="100%"
+            color={color || 'primary.400'}
+          />
+        )}
       </Link>
     </Box>
   )