+++ src/graphql/mutations/project.ts
@@ -2,6 +2,15 @@ import { gql } from '@apollo/client'
 
 import { FRAGMENT_PROJECT_REWARD_FOR_CREATE_UPDATE } from '../fragments/project'
 
+export const MUTATION_PROJECT_STATUS_UPDATE = gql`
+  mutation ProjectStatusUpdate($input: ProjectStatusUpdate!) {
+    projectStatusUpdate(input: $input) {
+      id
+      status
+    }
+  }
+`
+
 export const MUTATION_CREATE_PROJECT = gql`
   mutation CreateProject($input: CreateProjectInput!) {
     createProject(input: $input) {