+++ src/pages/projectCreate/ProjectCreationWalletConnectionForm.tsx
@@ -35,8 +35,10 @@ import {
   CreateWalletInput,
   LndNodeType,
   ProjectFragment,
+  ProjectStatus,
   useCreateWalletMutation,
   useLightningAddressVerifyLazyQuery,
+  useProjectStatusUpdateMutation,
   WalletResourceType,
 } from '../../types/generated/graphql'
 import { toInt, useNotification, validateEmail } from '../../utils'
@@ -186,6 +188,10 @@ export const ProjectCreationWalletConnectionForm = ({
     }
 
     if (connectionOption === ConnectionOption.LIGHTNING_ADDRESS) {
+      if (!lightningAddressFormValue) {
+        return null
+      }
+
       return {
         lightningAddressConnectionDetailsInput: {
           lightningAddress: lightningAddressFormValue,
@@ -210,7 +216,7 @@ export const ProjectCreationWalletConnectionForm = ({
   }, [connectionOption, lightningAddressFormValue, createWalletInput])
 
   const validateLightningAddress = async () => {
-    if (lightningAddressFormError === null) {
+    if (lightningAddressFormError === null && lightningAddressFormValue) {
       await evaluateLightningAddress()
     }
   }
@@ -247,6 +253,9 @@ export const ProjectCreationWalletConnectionForm = ({
     }
   }
 
+  const [updateStatus, { loading: isUpdateStatusLoading }] =
+    useProjectStatusUpdateMutation()
+
   const handleLaunch = async () => {
     await validateLightningAddress()
 
@@ -260,12 +269,19 @@ export const ProjectCreationWalletConnectionForm = ({
     }
 
     await createWallet({ variables: { input: createWalletInput } })
+    await updateStatus({
+      variables: {
+        input: { projectId: project.id, status: ProjectStatus.Active },
+      },
+    })
   }
 
   const validateLightningAddressFormat = async (lightningAddress: string) => {
-    if (lightningAddress.length === 0) {
-      setLightningAddressFormError(`Lightning Address can't be empty.`)
-    } else if (lightningAddress.endsWith('@geyser.fund')) {
+    if (!lightningAddress) {
+      return setLightningAddressFormError(null)
+    }
+
+    if (lightningAddress.endsWith('@geyser.fund')) {
       setLightningAddressFormError(
         `Custom Lightning Addresses can't end with "@geyser.fund".`,
       )
@@ -278,6 +294,11 @@ export const ProjectCreationWalletConnectionForm = ({
     }
   }
 
+  const isLoading =
+    isCreateWalletLoading ||
+    isUpdateStatusLoading ||
+    isEvaluatingLightningAddress
+
   if (isReadyForLaunch) {
     return (
       <ProjectCreateCompleted>
@@ -288,10 +309,10 @@ export const ProjectCreationWalletConnectionForm = ({
               w="full"
               leftIcon={<BiRocket />}
               onClick={onLaunchClick}
-              isLoading={isCreateWalletLoading}
+              isLoading={isLoading}
               disabled={
                 isSubmitEnabled === false ||
-                isEvaluatingLightningAddress ||
+                isLoading ||
                 Boolean(lightningAddressFormError)
               }
             >
@@ -302,7 +323,7 @@ export const ProjectCreationWalletConnectionForm = ({
             variant="secondary"
             w="full"
             onClick={onSaveDraftClick}
-            disabled={isCreateWalletLoading || isEvaluatingLightningAddress}
+            disabled={isLoading}
           >
             Save As Draft
           </Button>
@@ -423,11 +444,20 @@ export const ProjectCreationWalletConnectionForm = ({
             <WalletConnectionOptionInfoBox
               primaryText="Connect your node"
               promoText="No fee per transaction"
-              secondaryText="Connect your lightning node to receive incoming transactions directly. Don't have a node? You can create a cloud node with the recommended app."
+              secondaryText={
+                <span>
+                  Connect your lightning node to receive incoming transactions
+                  directly. Don&apos;t have a node? You can{' '}
+                  <Link href={VoltageExplainerPageForGeyserURL}>
+                    create a cloud node
+                  </Link>{' '}
+                  with the recommended app.
+                </span>
+              }
             >
               <HStack>
                 <Link isExternal href={VoltageExplainerPageForGeyserURL}>
-                  <Image src={VoltageUrl} />
+                  <Image maxWidth="8em" src={VoltageUrl} />
                 </Link>
               </HStack>
             </WalletConnectionOptionInfoBox>